+++ Src/Library/changelog.md
@@ -1,2 +1,5 @@
+### NEW
+- add `ProducesProblemFE()` extension method for `RouteHandlerBuilder` [#info](https://discord.com/channels/933662816458645504/1004762111546769498)
+
 ### IMPROVEMENTS
 - increase logging in validation schema processor #117